+++ packages/components/grid/src/column.ts
@@ -1,15 +1,17 @@
-import type { TemplateResult } from 'lit';
+import type { CSSResult, TemplateResult } from 'lit';
 import type { Grid } from './grid.js';
 import type { EventEmitter } from '@sl-design-system/shared';
 import { dasherize, event, getNameByPath, getValueByPath } from '@sl-design-system/shared';
 import { LitElement, html } from 'lit';
 import { property } from 'lit/decorators.js';
-import { ifDefined } from 'lit/directives/if-defined.js';
 
-export type GridColumnHeaderRenderer = () => TemplateResult;
+/** Custom renderer type for column headers. */
+export type GridColumnHeaderRenderer = () => string | undefined | TemplateResult;
 
-export type GridColumnDataRenderer<T> = (model: T) => TemplateResult;
+/** Custom renderer type for column cells. */
+export type GridColumnDataRenderer<T> = (model: T) => string | undefined | TemplateResult;
 
+/** Custom type for providing parts to a cell. */
 export type GridColumnParts<T> = (model: T) => string | undefined;
 
 export class GridColumn<T extends Record<string, unknown> = Record<string, unknown>> extends LitElement {
@@ -84,32 +86,43 @@ export class GridColumn<T extends Record<string, unknown> = Record<string, unkno
     super.connectedCallback();
 
     // If width and autoWidth are not set, set autoWidth to true
-    if (!this.width && this.autoWidth === undefined) {
+    if (this.width === undefined && this.autoWidth === undefined) {
       this.autoWidth = true;
     }
   }
 
   renderHeader(): TemplateResult {
-    return html`<th>${this.header ?? getNameByPath(this.path)}</th>`;
+    const parts = ['header', ...this.#getParts()];
+
+    return html`<th part=${parts.join(' ')}>${this.header ?? getNameByPath(this.path)}</th>`;
   }
 
   renderData(item: T): TemplateResult {
-    let parts;
-
-    if (typeof this.parts === 'string') {
-      parts = this.parts;
-    } else if (typeof this.parts === 'function') {
-      parts = this.parts(item);
-    }
-
-    if (this.path) {
-      parts = `${dasherize(this.path.replaceAll('.', '-'))} ${parts || ''}`.trim();
-    }
+    const parts = ['data', ...this.#getParts(item)];
 
     return html`
-      <td part=${ifDefined(parts)}>
+      <td part=${parts.join(' ')}>
         ${this.renderer ? this.renderer(item) : this.path ? getValueByPath(item, this.path) : 'No path set'}
       </td>
     `;
+  }
+
+  // eslint-disable-next-line @typescript-eslint/no-empty-function
+  renderStyles(): CSSResult | void {}
+
+  #getParts(item?: T): string[] {
+    let parts: string[] = [];
+
+    if (typeof this.parts === 'string') {
+      parts = this.parts.split(' ');
+    } else if (typeof this.parts === 'function' && item) {
+      parts = this.parts(item)?.split(' ') ?? [];
+    }
+
+    if (this.path) {
+      parts.push(dasherize(this.path.replaceAll('.', '-')));
+    }
+
+    return parts;
   }
 }