+++ packages/components/grid/src/grid.ts
@@ -13,7 +13,7 @@ import { virtualize } from '@lit-labs/virtualizer/virtualize.js';
 import { ScopedElementsMixin } from '@open-wc/scoped-elements';
 import { ArrayDataSource, SelectionController, event } from '@sl-design-system/shared';
 import { LitElement, html } from 'lit';
-import { property, state } from 'lit/decorators.js';
+import { property, query, state } from 'lit/decorators.js';
 import { classMap } from 'lit/directives/class-map.js';
 import styles from './grid.scss.js';
 import { GridColumn } from './column.js';
@@ -43,6 +43,9 @@ export class Grid<T extends Record<string, unknown> = Record<string, unknown>> e
   /** Flag for calculating the column widths only once. */
   #initialColumnWidthsCalculated = false;
 
+  /** Observe the grid width. */
+  #resizeObserver?: ResizeObserver;
+
   /** The sorters for this grid. */
   #sorters: GridSorter[] = [];
 
@@ -61,6 +64,12 @@ export class Grid<T extends Record<string, unknown> = Record<string, unknown>> e
   /** Provide your own implementation for getting the data. */
   @property({ attribute: false }) dataSource?: DataSource<T>;
 
+  /** The `<tbody>` element. */
+  @query('tbody') tbody!: HTMLTableSectionElement;
+
+  /** The `<thead>` element. */
+  @query('thead') thead!: HTMLTableSectionElement;
+
   /** An array of items to be displayed in the grid. */
   @property({ type: Array }) items?: T[];
 
@@ -75,6 +84,30 @@ export class Grid<T extends Record<string, unknown> = Record<string, unknown>> e
 
   /** Uses alternating background colors for the rows when set. */
   @property({ type: Boolean, reflect: true }) striped?: boolean;
+
+  override connectedCallback(): void {
+    super.connectedCallback();
+
+    this.#resizeObserver = new ResizeObserver(entries => {
+      const {
+        borderBoxSize: [{ inlineSize }]
+      } = entries[0];
+
+      this.style.setProperty('--sl-grid-width', `${inlineSize}px`);
+    });
+
+    this.#resizeObserver.observe(this);
+  }
+
+  override firstUpdated(): void {
+    this.tbody.addEventListener(
+      'scroll',
+      () => {
+        this.thead.scrollLeft = this.tbody.scrollLeft;
+      },
+      { passive: true }
+    );
+  }
 
   override willUpdate(changes: PropertyValues<this>): void {
     if (changes.has('items')) {
@@ -96,13 +129,20 @@ export class Grid<T extends Record<string, unknown> = Record<string, unknown>> e
     }
   }
 
+  override disconnectedCallback(): void {
+    this.#resizeObserver?.disconnect();
+    this.#resizeObserver = undefined;
+
+    super.disconnectedCallback();
+  }
+
   override render(): TemplateResult {
     return html`
       <slot @sl-column-update=${this.#onColumnUpdate} @slotchange=${this.#onSlotchange} style="display:none"></slot>
       <style>
         ${this.renderStyles()}
       </style>
-      <table>
+      <table part="table">
         <thead
           @sl-direction-change=${this.#onDirectionChange}
           @sl-filter-change=${this.#onFilterChange}
@@ -112,7 +152,7 @@ export class Grid<T extends Record<string, unknown> = Record<string, unknown>> e
         >
           ${this.renderHeader()}
         </thead>
-        <tbody @visibilityChanged=${this.#onVisibilityChanged}>
+        <tbody @visibilityChanged=${this.#onVisibilityChanged} part="tbody">
           ${virtualize({
             items: this.dataSource?.items,
             renderItem: (item, index) => this.renderItem(item, index)
@@ -131,8 +171,9 @@ export class Grid<T extends Record<string, unknown> = Record<string, unknown>> e
           return `
             thead tr:nth-child(${rowIndex + 1}) th:nth-child(${colIndex + 1}) {
               flex-grow: ${(col as GridColumnGroup<T>).columns.length};
+              inline-size: ${col.width || '100'}px;
               justify-content: ${col.align};
-              width: ${col.width || '100'}px;
+              ${col.renderStyles()?.toString() ?? ''}
             }
           `;
         });
@@ -141,16 +182,17 @@ export class Grid<T extends Record<string, unknown> = Record<string, unknown>> e
         return `
           :where(td, thead tr:last-of-type th):nth-child(${index + 1}) {
             flex-grow: ${col.grow};
+            inline-size: ${col.width || '100'}px;
             justify-content: ${col.align};
-            width: ${col.width || '100'}px;
             ${
               col.sticky
                 ? `
-                  left: ${this.#getStickyColumnOffset(index)}px;
+                  inset-inline-start: ${this.#getStickyColumnOffset(index)}px;
                   position: sticky;
                 `
                 : ''
             }
+            ${col.renderStyles()?.toString() ?? ''}
           }
         `;
       })}
@@ -217,6 +259,12 @@ export class Grid<T extends Record<string, unknown> = Record<string, unknown>> e
           return Math.max(acc, width);
         }, 0);
       });
+
+    // Since we set an explicit width for the `<thead>` and `<tbody>`, we also need
+    // to set an explicit with for all the `<tr>` elements. Otherwise, the sticky columns
+    // will not be sticky when you scroll horizontally.
+    const rowWidth = this.columns.reduce((acc, cur) => acc + (cur.width ?? 0), 0);
+    this.style.setProperty('--sl-grid-row-width', `${rowWidth}px`);
 
     this.requestUpdate('columns');
   }