+++ scripts/build-tokens.js
@@ -1,5 +1,6 @@
 import { promises as fs } from 'fs';
 import { join } from 'path';
+import { argv } from 'node:process';
 import sass from 'sass';
 import stylelint from 'stylelint';
 import figmaThemes from '../packages/tokens/src/$themes.json' assert { type: 'json' };
@@ -85,25 +86,31 @@ const buildTokens = async name => {
   });  
 };
 
-const buildAllTokens = async () => {
-  const themes = figmaThemes.map(({ name: nameVariant, selectedTokenSets }) => {
-    const [name, mode] = nameVariant.split('/');
-  
-    return {
-      name,
-      mode,
-      tokenSets: Object.entries(selectedTokenSets)
-        .filter(([_, enabled]) => enabled === 'enabled')
-        .filter(([name]) => name !== 'core' && !name.endsWith('/base'))
-        .map(([name]) => name)
-    };
-  });
+const buildAllTokens = async (themesToBuild) => {
+  const themes = figmaThemes
+    .filter(({ name }) => themesToBuild.length === 0 || themesToBuild.includes(name))
+    .map(({ name: nameVariant, selectedTokenSets }) => {
+      const [name, mode] = nameVariant.split('/');
+    
+      return {
+        name,
+        mode,
+        tokenSets: Object.entries(selectedTokenSets)
+          .filter(([_, enabled]) => enabled === 'enabled')
+          .filter(([name]) => name !== 'core' && !name.endsWith('/base'))
+          .map(([name]) => name)
+      };
+    });
   
   figmaThemes
     .reduce((acc, { name }) => {
-      const id = name.split('/')[0];
+      if (themesToBuild.length === 0 || themesToBuild.includes(name)) {
+        const id = name.split('/')[0];
   
-      return acc.includes(id) ? acc : [...acc, id];
+        return acc.includes(id) ? acc : [...acc, id];
+      } else {
+        return acc;
+      }
     }, [])
     .forEach(id => {
       themes.push({
@@ -118,4 +125,4 @@ const buildAllTokens = async () => {
   themes.map(({ name }) => buildTokens(name));
 };
 
-buildAllTokens();
+buildAllTokens(argv.slice(2));