+++ packages/themes/magister/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @sanomalearning/slds-magister
 
+## 0.0.3
+
+### Patch Changes
+
+- [#407](https://github.com/sl-design-system/components/pull/407) [`3da64be`](https://github.com/sl-design-system/components/commit/3da64be6a356b05ce8ff087f5ef9af15e85e7090) - Checkbox and radio button tokens added
+
+- [#426](https://github.com/sl-design-system/components/pull/426) [`41663ef`](https://github.com/sl-design-system/components/commit/41663efecec6241a17aa262b4b8c7d52609f16dd) - Added a way to have custom icons on core SLDS level, so we can serve those along with FontAwesome tokens to all themes
+
+- [#430](https://github.com/sl-design-system/components/pull/430) [`d66c3c3`](https://github.com/sl-design-system/components/commit/d66c3c37b73ebaf34a336f258b9d59ddd21152fc) - Change non-button components to have as 1px border instead of 2px
+
+- Updated dependencies [[`d0eae48`](https://github.com/sl-design-system/components/commit/d0eae48a112ec6c096ca6f3804cb248a390f04c8)]:
+  - @sl-design-system/icon@0.0.3
+
 ## 0.0.2
 
 ### Patch Changes